feat: crear ProductTable.tsx como componente independiente
- Extraer tabla de productos a componente reutilizable ProductTable.tsx
- Maneja su propia query y mutations internamente
- Simplificar AdminDashboard.tsx usando el nuevo componente
- Eliminar código redundante de gestión de productos en dashboard
